Time: 22 ms (56.71%) | Memory: 17.4 MB (54.86%) - LeetSync

+++ 18-4sum/4sum.cpp
@@ -15,7 +15,7 @@ public:
                 int end = n - 1;
 
                 while(start < end){
-                    long long sum = (long)(long)nums[i] + nums[j] + nums[start] + nums[end];
+                    long long sum = (long long)nums[i] + nums[j] + nums[start] + nums[end];
                     if(sum < target){
                         start++;
                     } else if(sum > target){
